Refactor EventQuery to Event.Query type
Moved EventQuery to Event.Query as a nested struct for better organization and namespacing. Updated all references to use Event.Query instead of EventQuery, and removed the old EventQuery file.
